+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.34.3
+## 3.4.35.0
 
 ### New Features
+- Add api key setting for Modrinth to allow accessing private packs
 
 ### Fixes
 - Issue with servers button not disabling when cannot create servers on Create Pack tab